Features Added------------------------------------------------------- Customers may update account information, or delete his/her account.
You should only allow the staff to delete a product if no one is interested in buying it in the past month.

When a product is low on stock, staff should be able to get an alert to get more from the supplier.

Features Needed--------------------------------------------------------

---Can add button for logged in users when viewing item in store-----
Customers may add multiple products to one order and make a payment with the exact amount of the order. Only registered user can make orders.

---Must Create Assertion For This-------
No overselling is allowed. Can't allow user to order more than what is in stock.